+++ store/primary/src/store/orchestrator_actions.rs
@@ -1,5 +1,4 @@
 use opentracingrust::SpanContext;
-use uuid::Uuid;
 
 use replicante_models_core::actions::orchestrator::OrchestratorActionSyncSummary;
 use replicante_models_core::api::orchestrator_action::OrchestratorActionSummary;
@@ -20,22 +19,6 @@ impl OrchestratorActions {
         attrs: OrchestratorActionsAttributes,
     ) -> OrchestratorActions {
         OrchestratorActions { actions, attrs }
-    }
-
-    /// Approve a PENDING_APPROVE orchestrator action for scheduling and execution.
-    pub fn approve<S>(&self, action_id: Uuid, span: S) -> Result<()>
-    where
-        S: Into<Option<SpanContext>>,
-    {
-        self.actions.approve(&self.attrs, action_id, span.into())
-    }
-
-    /// Disapprove a PENDING_APPROVE orchestrator action so it won't be scheduled.
-    pub fn disapprove<S>(&self, action_id: Uuid, span: S) -> Result<()>
-    where
-        S: Into<Option<SpanContext>>,
-    {
-        self.actions.disapprove(&self.attrs, action_id, span.into())
     }
 
     /// Iterate over orchestrator actions for a cluster.