+++ CHANGELOG.md
@@ -7,13 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
+- Emit `ACTION_CHANGED` events on action approve and disapprove.
+- Emit `ACTION_CHANGED` events when node actions fail to schedule.
+- Emit `ACTION_ORCHESTRATOR_CHANGED` events on orchestrator approve and disapprove.
 - Errors during node sync are logged as info instead of debug.
 - Node action approve and disapprove endpoints reject requests for actions not pending approval.
 - Node action approve and disapprove endpoints reject requests for missing actions.
 
 ### Fix
-- Emit `ACTION_CHANGED` events on action approve and disapprove.
-- Emit `ACTION_CHANGED` events when node actions fail to schedule.
 - Fixed loading node and orchestrator actions from the database.
 
 ## [0.7.1] - 2022-09-14